pyproject.toml
[project]
name = "bqwizard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Rolando Franqui", email = "rolandofranqui@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "google-cloud-bigquery>=3.30.0",
    "platformdirs>=4.3.7",
    "pydantic-settings>=2.8.1",
    "tabulate>=0.9.0",
]

[project.scripts]
bqwizard = "bqwizard.main:cli"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bqwizard"]
